[ADD] Simplify `DirectionalDerivatives` API (#17)
Exotic features, like using different GGNs to compute directions and
directional curvatures, as well as full control of which intermediate buffers
to keep, have been deprecated in favor of a simpler API.

- Remove Newton step computation for now as it was internally relying on
 `DirectionalDerivatives`
- Remove many utilities and associated tests from the exotic features
- Forbid duplicate indices in `subsampling`
- Always delete intermediate buffers other than the target quantities

---

* [REF] Rename `GramComputations` → `DirectionalDerivativesComputation`

* [DEL] Remove `compute_gammas` and `keep_gammas` arguments

Always compute and keep first-order directional derivatives.

* [DEL] Remove `compute_lambdas` and `keep_lambdas` arguments

* [DEL] Remove Newton step and Newton optimizer

Currently, the Newton step depends on the directional derivatives.
I will be easier to clean up the latter's API before reactivating the
Newton step feature.

* [DEL] Remove `keep_gram_mat` argument

Always delete the Gram matrix.

* [DEL] Remove `keep_gram_evecs` argument

Always remove the Gram matrix eigenvectors.

* [DEL] Remove `keep_gram_evals` argument

Always delete the Gram matrix eigenvalues.

* [DEL] Remove `keep_batch_size` argument

Always delete batch size.

* [DEL] Remove `keep_backpack_buffers` argument

Always delete BackPACK buffers.

* [DEL] Remove `param_groups` from `get_extensions`

* [ADD] Simplify `DirectionalDerivativesComputation` API and tests

* [ADD] Forbid `subsampling` with repeated indices (#16)

* [REF] Rename file containing argument checks

* [ADD] Forbid sub-sampling with repeated indices


* [ADD] Simplify internals of `DirectionalDerivativesComputation`

Forbid duplicates in subsampling, share subsampling for directions and
directional curvatures.
